subject_selection/requirements.txt
bs4
selenium
tranco
selenium-wire
pymongo
python-dotenv
lxml
